+++ HISTORY.md
@@ -1,4 +1,7 @@
 # Changelog
 
-## Version 0.1.0a1
-* First development version
+## Version <=0.3.0a1
+* Legacy implementations.
+
+## Version 0.3.0a1
+* Reworked to support concepts API and drop own FCA implementation.